Rename the files into new structure and format
