--- src/Kulku.Web/Kulku.Web.Admin/Components/Developer/Components/CrmCompanyEditModal.razor
@@ -1,106 +1,0 @@
-@using Kulku.Application.Network.Models
-@using Kulku.Domain.Network
-
-<ModalDialog Visible="@IsVisible" Title="@ModalTitle"
- Size="ModalSize.Large"
- OnClose="OnCancel">
-<ChildContent>
-<EditForm Model="@Form" OnValidSubmit="HandleSave">
-<DataAnnotationsValidator />
-
-@if (Mode == ModalMode.Create)
-{
-<div class="mb-3">
-<div class="btn-group btn-group-sm w-100 mb-2" role="group">
-<button type="button"
-class="btn @(!Form.IsNewCompany ? "btn-primary" : "btn-outline-primary")"
-@onclick="() => Form.IsNewCompany = false">
-Pick existing
-</button>
-<button type="button"
-class="btn @(Form.IsNewCompany ? "btn-primary" : "btn-outline-primary")"
-@onclick="() => Form.IsNewCompany = true">
-Create new
-</button>
-</div>
-
-@if (Form.IsNewCompany)
-{
-<div class="mb-2">
-<label class="form-label">Company Name <span class="text-danger">*</span></label>
-<InputText class="form-control" @bind-Value="Form.NewCompanyName" placeholder="Company name" />
-<ValidationMessage For="@(() => Form.NewCompanyName)" />
-</div>
-<div class="row g-2">
-<div class="col-md-6">
-<label class="form-label">Website</label>
-<InputText class="form-control" @bind-Value="Form.NewCompanyWebsite" placeholder="https://..." />
-</div>
-<div class="col-md-6">
-<label class="form-label">Region</label>
-<InputText class="form-control" @bind-Value="Form.NewCompanyRegion" placeholder="e.g. Helsinki, Finland" />
-</div>
-</div>
-}
-else
-{
-<label class="form-label">Company</label>
-@if (AvailableCompanies.Count == 0)
-{
-<div class="text-muted small">All companies are already enrolled in the network.</div>
-}
-else
-{
-<InputSelect class="form-select" @bind-Value="Form.SelectedCompanyId">
-<option value="@Guid.Empty">— Select a company —</option>
-@foreach (var c in AvailableCompanies.OrderBy(c => c.Name))
-{
-<option value="@c.CompanyId">@c.Name @(c.Region is not null ? $"({c.Region})" : "")</option>
-}
-</InputSelect>
-<ValidationMessage For="@(() => Form.SelectedCompanyId)" />
-}
-}
-</div>
-}
-
-<div class="mb-3">
-<label class="form-label">Stage</label>
-<InputSelect class="form-select" @bind-Value="Form.Stage">
-@foreach (var s in Enum.GetValues<CompanyStage>())
-{
-<option value="@s">@CrmCompanyCard.StageLabel(s)</option>
-}
-</InputSelect>
-</div>
-
-<div class="mb-3">
-<label class="form-label">Notes</label>
-<InputTextArea class="form-control" rows="3" @bind-Value="Form.Notes" />
-</div>
-
-<div class="mb-3">
-<label class="form-label">Categories</label>
-<div class="d-flex flex-wrap gap-2">
-@foreach (var cat in Categories.OrderBy(c => c.Name))
-{
-<div class="form-check">
-<input class="form-check-input" type="checkbox"
-   checked="@Form.CategoryIds.Contains(cat.Id)"
-   @onchange="e => ToggleCategory(cat.Id, (bool)e.Value!)" />
-<label class="form-check-label">@cat.Name</label>
-</div>
-}
-</div>
-</div>
-
-<div class="d-flex gap-2">
-<button type="submit" class="btn btn-primary"
-disabled="@IsSubmitDisabled">
-@SubmitLabel
-</button>
-<button type="button" class="btn btn-outline-secondary" @onclick="OnCancel">Cancel</button>
-</div>
-</EditForm>
-</ChildContent>
-</ModalDialog>
