--- src/Kulku.Web/Kulku.Web.Admin/Components/Developer/Components/RelationshipRow.razor
@@ -1,124 +1,0 @@
-@using Kulku.Domain.Network
-@using Kulku.Web.Admin.Components.Developer
-
-<div class="card mb-2 border-start border-3 @FreshnessClass">
-<div class="card-body py-2 px-3">
-<div class="d-flex justify-content-between align-items-start">
-<div class="flex-grow-1 me-2">
-@* Row 1: Name + stage + region *@
-<div class="d-flex align-items-center gap-2 mb-1">
-<h6 class="mb-0">
-<a href="/developer/companies/@Relationship.CompanyId" class="text-decoration-none">
-@Relationship.CompanyName
-</a>
-</h6>
-<span class="badge @CrmCompanyCard.StageBadgeClass(Relationship.Stage)">
-@CrmCompanyCard.StageLabel(Relationship.Stage)
-</span>
-@if (!string.IsNullOrWhiteSpace(Relationship.Region))
-{
-<small class="text-muted">@Relationship.Region</small>
-}
-</div>
-
-@* Row 2: Categories *@
-@if (Relationship.Categories.Count > 0)
-{
-<div class="d-flex flex-wrap gap-1 mb-1">
-@foreach (var cat in Relationship.Categories)
-{
-<span class="badge text-bg-@(cat.ColorToken ?? "secondary") bg-opacity-50 fw-normal"
-  style="font-size: 0.75rem;">@cat.Name</span>
-}
-</div>
-}
-
-@* Row 3: Key contact + relationship age *@
-<div class="d-flex flex-wrap align-items-center gap-3 mb-1">
-@if (!string.IsNullOrWhiteSpace(Relationship.PrimaryContactName))
-{
-<small class="text-muted">
-<Icon Kind="IconKind.PersonLinesFill" Size="12" CssClass="me-1" />@Relationship.PrimaryContactName
-@if (!string.IsNullOrWhiteSpace(Relationship.PrimaryContactTitle))
-{
-<span class="opacity-75"> · @Relationship.PrimaryContactTitle</span>
-}
-</small>
-}
-@if (Relationship.FirstInteractionDate.HasValue)
-{
-<small class="text-muted opacity-75">
-Since @Relationship.FirstInteractionDate.Value.ToString("MMM yyyy")
-</small>
-}
-</div>
-
-@* Row 4: Last touch *@
-@if (Relationship.LastInteractionDate.HasValue)
-{
-<div class="d-flex align-items-center gap-2">
-<small class="text-muted fw-semibold">
-Last touch: @RelativeTime(Relationship.LastInteractionDate.Value)
-</small>
-@if (Relationship.LastChannel.HasValue)
-{
-<span class="badge text-bg-light border" style="font-size: 0.7rem;">
-@ChannelLabel(Relationship.LastChannel.Value)
-</span>
-}
-@if (!string.IsNullOrWhiteSpace(Relationship.LastSummary))
-{
-<small class="text-muted fst-italic">
-"@Truncate(Relationship.LastSummary, 80)"
-</small>
-}
-</div>
-}
-else
-{
-<small class="text-muted fst-italic">No interactions recorded yet</small>
-}
-
-@* Row 5: Next step (if any) *@
-@if (!string.IsNullOrWhiteSpace(Relationship.NextAction))
-{
-<div class="d-flex align-items-center gap-2 mt-1">
-<small class="text-muted">Next:</small>
-<small>@Truncate(Relationship.NextAction, 60)</small>
-@if (Relationship.NextActionDue.HasValue)
-{
-<span class="badge @DueBadge(Relationship.NextActionDue.Value)" style="font-size: 0.7rem;">
-@Relationship.NextActionDue.Value.ToString("yyyy-MM-dd")
-</span>
-}
-</div>
-}
-</div>
-
-@* Right side: health indicator *@
-<div class="d-flex flex-column align-items-center gap-1 flex-shrink-0">
-@if (Relationship.DaysSinceLastTouch.HasValue)
-{
-<div class="text-center">
-<div class="h5 mb-0 @HealthTextClass">@Relationship.DaysSinceLastTouch.Value</div>
-<small class="text-muted" style="font-size: 0.65rem;">days</small>
-</div>
-}
-else
-{
-<small class="text-muted">—</small>
-}
-</div>
-</div>
-</div>
-</div>
-
-@code {
-private string HealthTextClass => Relationship.Health switch
-{
-RelationshipHealth.Fresh => "text-success",
-RelationshipHealth.Cooling => "text-warning",
-RelationshipHealth.Cold => "text-danger",
-_ => "text-muted",
-};
-}
