--- src/Kulku.Web/Kulku.Web.Admin/Components/Developer/Pages/CompanyDetailPage.razor
@@ -1,236 +1,0 @@
-@page "/developer/companies/{CompanyId:guid}"
-@using Kulku.Application.Network.Models
-@using Kulku.Domain.Network
-@using Kulku.Web.Admin.Components.Developer.Components
-
-@attribute [Authorize]
-
-@if (_company is null)
-{
-<PageTitle>Company not found</PageTitle>
-<div class="container-fluid py-4">
-<div class="alert alert-warning">Company not found.</div>
-<NavLink class="btn btn-sm btn-outline-secondary" href="developer/companies">Back to companies</NavLink>
-</div>
-}
-else
-{
-<PageTitle>@_company.Name</PageTitle>
-
-<div class="container-fluid py-4">
-@* ── Header ── *@
-<div class="d-flex flex-wrap justify-content-between align-items-start gap-2 mb-4">
-<div>
-<div class="d-flex align-items-center gap-2 mb-1">
-<NavLink class="btn btn-sm btn-outline-secondary" href="developer/companies">
-<span style="display:inline-block;transform:rotate(180deg)"><Icon Kind="IconKind.ChevronRight" Size="14" /></span> Back
-</NavLink>
-</div>
-<h1 class="h3 mb-1">@_company.Name</h1>
-<div class="d-flex align-items-center gap-2">
-<span class="badge @CrmCompanyCard.StageBadgeClass(_company.Stage)">
-@CrmCompanyCard.StageLabel(_company.Stage)
-</span>
-@if (!string.IsNullOrWhiteSpace(_company.Region))
-{
-<span class="text-muted">@_company.Region</span>
-}
-@if (!string.IsNullOrWhiteSpace(_company.Website))
-{
-<a href="@_company.Website" target="_blank" class="text-decoration-none small">
-@_company.Website
-</a>
-}
-</div>
-@if (_company.Categories.Count > 0)
-{
-<div class="d-flex flex-wrap gap-1 mt-2">
-@foreach (var cat in _company.Categories)
-{
-<span class="badge text-bg-@(cat.ColorToken ?? "secondary") bg-opacity-50 fw-normal">
-@cat.Name
-</span>
-}
-</div>
-}
-</div>
-
-<div class="d-flex flex-wrap gap-2">
-<button class="btn btn-sm btn-outline-primary" @onclick="HandleEditCompany">
-<Icon Kind="IconKind.Pencil" Size="14" CssClass="me-1" />Edit Company
-</button>
-<NavLink class="btn btn-sm btn-primary" href="developer/loginteraction">
-<Icon Kind="IconKind.Plus" Size="14" CssClass="me-1" />Log Interaction
-</NavLink>
-</div>
-</div>
-
-@* ── Stage selector ── *@
-<div class="card shadow-sm mb-4">
-<div class="card-body py-2">
-<div class="d-flex align-items-center gap-2">
-<span class="small text-muted me-2">Stage:</span>
-@foreach (var stage in Enum.GetValues<CompanyStage>())
-{
-<button class="btn btn-sm @(stage == _company.Stage ? CrmCompanyCard.StageBadgeClass(stage).Replace("text-bg-", "btn-") : "btn-outline-secondary")"
-@onclick="() => ChangeStage(stage)">
-@CrmCompanyCard.StageLabel(stage)
-</button>
-}
-</div>
-</div>
-</div>
-
-@* ── Notes ── *@
-@if (!string.IsNullOrWhiteSpace(_company.Notes))
-{
-<div class="card shadow-sm mb-4">
-<div class="card-header py-2">
-<strong>Notes</strong>
-</div>
-<div class="card-body py-2">
-<p class="mb-0 text-muted">@_company.Notes</p>
-</div>
-</div>
-}
-
-<div class="row g-4">
-@* ── Contacts ── *@
-<div class="col-12 col-xl-4">
-<div class="card shadow-sm h-100">
-<div class="card-header py-2 d-flex justify-content-between align-items-center">
-<div class="d-flex align-items-center gap-2">
-<strong>Contacts</strong>
-<span class="badge bg-secondary">@_contacts.Count</span>
-</div>
-<button class="btn btn-sm btn-outline-success" @onclick="HandleCreateContact">
-<Icon Kind="IconKind.Plus" Size="14" />
-</button>
-</div>
-<div class="card-body p-2">
-@if (_contacts.Count == 0)
-{
-<div class="text-muted small text-center py-3">No contacts yet.</div>
-}
-else
-{
-@foreach (var contact in _contacts)
-{
-<div class="card mb-2 border-0 bg-body-secondary">
-<div class="card-body py-2 px-3">
-<div class="d-flex justify-content-between align-items-start">
-<div>
-<div class="fw-semibold small">@(contact.PersonName ?? "(unnamed)")</div>
-@if (!string.IsNullOrWhiteSpace(contact.Title))
-{
-<small class="text-muted">@contact.Title</small>
-}
-<div class="d-flex flex-wrap gap-2 mt-1">
-@if (!string.IsNullOrWhiteSpace(contact.Email))
-{
-<small class="text-muted">📧 @contact.Email</small>
-}
-@if (!string.IsNullOrWhiteSpace(contact.Phone))
-{
-<small class="text-muted">📱 @contact.Phone</small>
-}
-</div>
-@if (!string.IsNullOrWhiteSpace(contact.LinkedInUrl))
-{
-<div class="mt-1">
-<a href="@contact.LinkedInUrl" target="_blank"
-   class="text-decoration-none small">LinkedIn</a>
-</div>
-}
-</div>
-
-<div class="d-flex gap-1">
-<button class="btn btn-sm btn-outline-primary" title="Edit"
-@onclick="() => HandleEditContact(contact)">
-<Icon Kind="IconKind.Pencil" Size="12" />
-</button>
-<div class="dropdown">
-<button class="btn btn-sm btn-outline-secondary dropdown-toggle" title="Move to…"
-data-bs-toggle="dropdown" data-bs-auto-close="true">
-<Icon Kind="IconKind.PersonLinesFill" Size="12" />
-</button>
-<ul class="dropdown-menu dropdown-menu-end">
-<li><h6 class="dropdown-header">Move to…</h6></li>
-@foreach (var c in _enrolledCompanies.Where(c => c.CompanyId != CompanyId).OrderBy(c => c.Name))
-{
-<li>
-<button class="dropdown-item small"
-@onclick="() => HandleMoveContact(contact.Id, c.CompanyId)">
-@c.Name
-</button>
-</li>
-}
-<li><hr class="dropdown-divider" /></li>
-<li>
-<button class="dropdown-item small text-muted"
-@onclick="() => HandleUnaffiliateContact(contact.Id)">
-Remove affiliation (personal)
-</button>
-</li>
-</ul>
-</div>
-<button class="btn btn-sm btn-outline-danger" title="Delete"
-@onclick="() => HandleDeleteContact(contact.Id)">
-<Icon Kind="IconKind.Trash" Size="12" />
-</button>
-</div>
-</div>
-</div>
-</div>
-}
-}
-</div>
-</div>
-</div>
-
-@* ── Interaction Timeline ── *@
-<div class="col-12 col-xl-8">
-<div class="card shadow-sm h-100">
-<div class="card-header py-2 d-flex justify-content-between align-items-center">
-<div class="d-flex align-items-center gap-2">
-<strong>Interaction Timeline</strong>
-<span class="badge bg-secondary">@_interactions.Count</span>
-</div>
-</div>
-<div class="card-body p-2">
-@if (_interactions.Count == 0)
-{
-<div class="text-muted small text-center py-3">
-No interactions recorded.
-<NavLink href="developer/loginteraction" class="text-decoration-none">Log one now.</NavLink>
-</div>
-}
-else
-{
-@foreach (var interaction in _interactions)
-{
-<InteractionCard Interaction="@interaction"
- ShowCompanyName="false"
- OnDelete="HandleDeleteInteraction" />
-}
-}
-</div>
-</div>
-</div>
-</div>
-</div>
-
-@* ── Modals ── *@
-<CrmCompanyEditModal @ref="_companyModal"
- IsVisible="@(_companyEditVisible)"
- Mode="ModalMode.Edit"
- Categories="@_categories"
- OnSave="HandleSaveCompany"
- OnCancel="CloseCompanyEditor" />
-
-<CrmContactEditModal @ref="_contactModal"
- IsVisible="@(_contactMode is not null)"
- Mode="@(_contactMode ?? ModalMode.Create)"
- OnSave="HandleSaveContact"
- OnCancel="CloseContactEditor" />
-}
