--- src/Kulku.Web/Kulku.Web.Admin/Components/Developer/Pages/Index.razor
@@ -1,248 +1,0 @@
-@page "/developer/"
-@using Kulku.Application.Network.Models
-@using Kulku.Domain.Network
-@using Kulku.Web.Admin.Components.Developer.Components
-
-@attribute [Authorize]
-
-<PageTitle>Network Overview</PageTitle>
-
-<div class="container-fluid py-4">
-
-<div class="d-flex flex-wrap justify-content-between align-items-start gap-2 mb-4">
-<div>
-<h1 class="h3 mb-1">Network Overview</h1>
-<div class="text-muted">
-Your professional network — who you know, what happened, and where to reconnect.
-</div>
-</div>
-
-<div class="d-flex flex-wrap gap-2">
-<NavLink class="btn btn-sm btn-outline-secondary" href="developer/companies">
-<Icon Kind="IconKind.Building" Size="14" CssClass="me-1" />Companies
-</NavLink>
-<NavLink class="btn btn-sm btn-outline-secondary" href="developer/activity">
-<Icon Kind="IconKind.ChatDots" Size="14" CssClass="me-1" />Activity log
-</NavLink>
-<NavLink class="btn btn-sm btn-primary" href="developer/loginteraction">
-<Icon Kind="IconKind.Plus" Size="14" CssClass="me-1" />Log interaction
-</NavLink>
-</div>
-</div>
-
-@* ── Network Pulse ── *@
-<div class="card shadow-sm mb-4">
-<div class="card-body py-3">
-<div class="d-flex flex-wrap align-items-center gap-4">
-<div class="d-flex align-items-baseline gap-2">
-<span class="h4 mb-0">@_totalRelationships</span>
-<span class="text-muted small">relationships</span>
-</div>
-
-<div class="vr d-none d-md-block"></div>
-
-<div class="d-flex align-items-baseline gap-2">
-<span class="h4 mb-0 text-success">@_activeCount</span>
-<span class="text-muted small">active <span class="opacity-50">(12mo)</span></span>
-</div>
-
-<div class="vr d-none d-md-block"></div>
-
-<div class="d-flex align-items-baseline gap-2">
-<span class="h4 mb-0 @(_goingColdCount > 0 ? "text-danger" : "text-muted")">@_goingColdCount</span>
-<span class="text-muted small">going cold</span>
-</div>
-
-<div class="vr d-none d-md-block"></div>
-
-<div class="d-flex align-items-baseline gap-2">
-<span class="h4 mb-0">@_followUpsDue</span>
-<span class="text-muted small">follow-ups</span>
-@if (_overdueCount > 0)
-{
-<span class="badge text-bg-danger">@_overdueCount overdue</span>
-}
-</div>
-</div>
-</div>
-</div>
-
-@* ── Pipeline summary ── *@
-<div class="card shadow-sm mb-4">
-<div class="card-body py-2">
-<div class="d-flex align-items-center gap-3 flex-wrap">
-<span class="small text-muted">Pipeline:</span>
-@foreach (var stage in Enum.GetValues<CompanyStage>())
-{
-var count = _companies.Count(c => c.Stage == stage);
-<NavLink href="developer/companies" class="text-decoration-none">
-<span class="badge @CrmCompanyCard.StageBadgeClass(stage)">
-@CrmCompanyCard.StageLabel(stage) @count
-</span>
-</NavLink>
-}
-</div>
-</div>
-</div>
-
-@* ── Relationship Directory + Attention Needed ── *@
-<div class="row g-3 mb-4">
-<div class="col-12 col-xl-8">
-<div class="card shadow-sm h-100">
-<div class="card-header d-flex flex-wrap justify-content-between align-items-center gap-2">
-<div>
-<strong>Relationship Directory</strong>
-<div class="text-muted small">Every tracked company — coldest relationships first.</div>
-</div>
-</div>
-
-<div class="card-body pb-2 pt-3">
-@* Filters *@
-<div class="row g-2 mb-3">
-<div class="col-md-5">
-<input type="text" class="form-control form-control-sm" placeholder="Search name, region, contact…"
-   @bind="_searchText" @bind:event="oninput" />
-</div>
-<div class="col-md-3">
-<select class="form-select form-select-sm" @bind="_filterStage">
-<option value="">All stages</option>
-@foreach (var s in Enum.GetValues<CompanyStage>())
-{
-<option value="@s">@CrmCompanyCard.StageLabel(s)</option>
-}
-</select>
-</div>
-</div>
-
-@{
-var filtered = FilteredRelationships;
-}
-
-@if (filtered.Count == 0)
-{
-<div class="text-muted text-center py-4">
-@(_relationships.Count == 0
-? "No companies tracked yet. Enroll companies from the Companies page."
-: "No relationships match the current filters.")
-</div>
-}
-else
-{
-@foreach (var r in filtered)
-{
-<RelationshipRow Relationship="@r" />
-}
-}
-</div>
-
-<div class="card-footer text-muted small">
-Left border color: <span class="text-success">●</span> fresh (&lt;3mo)
-<span class="text-warning ms-2">●</span> cooling (3–6mo)
-<span class="text-danger ms-2">●</span> cold (&gt;6mo)
-<span class="text-secondary ms-2">●</span> no history
-</div>
-</div>
-</div>
-
-<div class="col-12 col-xl-4">
-<div class="card shadow-sm h-100">
-<div class="card-header">
-<strong>Attention Needed</strong>
-<div class="text-muted small">Overdue actions and relationships going cold.</div>
-</div>
-
-<div class="card-body">
-@if (_overdueActions.Count == 0 && _coldRelationships.Count == 0)
-{
-<div class="text-center py-4">
-<div class="text-success mb-2">
-<Icon Kind="IconKind.Check" Size="24" />
-</div>
-<div class="text-muted">All relationships are on track.</div>
-</div>
-}
-else
-{
-@if (_overdueActions.Count > 0)
-{
-<div class="mb-3">
-<div class="d-flex align-items-center gap-2 mb-2">
-<span class="badge text-bg-danger">Overdue</span>
-<small class="text-muted">@_overdueActions.Count action(s)</small>
-</div>
-<div class="list-group list-group-flush">
-@foreach (var item in _overdueActions)
-{
-<a href="/developer/companies/@item.CompanyId" class="list-group-item list-group-item-action px-0">
-<div class="d-flex justify-content-between align-items-start">
-<div class="fw-semibold small">@item.CompanyName</div>
-<span class="badge text-bg-danger">@item.DaysCount d overdue</span>
-</div>
-<div class="text-muted small">@item.Detail</div>
-</a>
-}
-</div>
-</div>
-}
-
-@if (_coldRelationships.Count > 0)
-{
-<div>
-<div class="d-flex align-items-center gap-2 mb-2">
-<span class="badge text-bg-warning">Going cold</span>
-<small class="text-muted">@_coldRelationships.Count relationship(s)</small>
-</div>
-<div class="list-group list-group-flush">
-@foreach (var item in _coldRelationships)
-{
-<a href="/developer/companies/@item.CompanyId" class="list-group-item list-group-item-action px-0">
-<div class="fw-semibold small">@item.CompanyName</div>
-<div class="text-muted small">@item.Detail</div>
-<small class="text-muted opacity-75">Consider checking in</small>
-</a>
-}
-</div>
-</div>
-}
-}
-</div>
-</div>
-</div>
-</div>
-
-@* ── Recent Activity ── *@
-<div class="row g-3">
-<div class="col-12">
-<div class="card shadow-sm">
-<div class="card-header d-flex flex-wrap justify-content-between align-items-center gap-2">
-<div>
-<strong>Latest Activity</strong>
-<div class="text-muted small">Your interaction journal.</div>
-</div>
-<NavLink href="developer/activity" class="btn btn-sm btn-outline-secondary text-decoration-none">
-View all
-</NavLink>
-</div>
-
-<div class="card-body p-2">
-@if (_streamInteractions.Count == 0)
-{
-<div class="p-3 text-muted">No interactions recorded yet.</div>
-}
-else
-{
-@foreach (var interaction in _streamInteractions)
-{
-<InteractionCard Interaction="@interaction"
- CompanyLink="@($"/developer/companies/{interaction.CompanyId}")" />
-}
-}
-</div>
-
-<div class="card-footer text-muted small">
-Showing up to 10 most recent. <a href="/developer/activity" class="text-decoration-none">View full history →</a>
-</div>
-</div>
-</div>
-</div>
-</div>
