+++ src/Kulku.Web/Kulku.Web.Admin/Components/Layout/MainLayout.razor
@@ -1,19 +1,19 @@
 @inherits LayoutComponentBase
 
 <div class="page">
-  <aside>
-    <NavMenu />
-  </aside>
+	<aside>
+		<NavMenu />
+	</aside>
 
-  <main>
-    <article class="content px-4">
-      @Body
-    </article>
-  </main>
+	<main>
+		<article class="content px-4">
+			@Body
+		</article>
+	</main>
 </div>
 
 <div id="blazor-error-ui" data-nosnippet>
-  An unhandled error has occurred.
-  <a href="." class="reload">Reload</a>
-  <span class="dismiss">🗙</span>
+	An unhandled error has occurred.
+	<a href="." class="reload">Reload</a>
+	<span class="dismiss">🗙</span>
 </div>
