+++ src/Kulku.Web/Kulku.Web.Admin/Components/Layout/NavMenu.razor
@@ -37,18 +37,18 @@
 
 		<AuthorizeView>
 			<li class="nav-item">
-				<NavLink class="nav-link text-white kulku-navlink" href="developer"
-						 data-bs-toggle="@(IsCollapsed ? "tooltip" : null)" data-bs-placement="right" title="@(IsCollapsed ? "Developer" : null)">
+				<NavLink class="nav-link text-white kulku-navlink" href="network"
+						 data-bs-toggle="@(IsCollapsed ? "tooltip" : null)" data-bs-placement="right" title="@(IsCollapsed ? "Network" : null)">
 					<Icon Kind="IconKind.Lightbulb" Size="20" />
-					<span class="kulku-link-text">Developer</span>
+					<span class="kulku-link-text">Network</span>
 				</NavLink>
 
 				@if (!IsCollapsed)
 				{
 					<ul class="nav flex-column kulku-subnav">
-						<li><NavLink class="nav-link kulku-sublink" href="developer" Match="NavLinkMatch.All"><Icon Kind="IconKind.Speedometer" Size="14" /> Dashboard</NavLink></li>
-						<li><NavLink class="nav-link kulku-sublink" href="developer/companies"><Icon Kind="IconKind.Building" Size="14" /> Companies</NavLink></li>
-						<li><NavLink class="nav-link kulku-sublink" href="developer/loginteraction"><Icon Kind="IconKind.ChatDots" Size="14" /> Log Interaction</NavLink></li>
+						<li><NavLink class="nav-link kulku-sublink" href="network" Match="NavLinkMatch.All"><Icon Kind="IconKind.Speedometer" Size="14" /> Dashboard</NavLink></li>
+						<li><NavLink class="nav-link kulku-sublink" href="network/companies"><Icon Kind="IconKind.Building" Size="14" /> Companies</NavLink></li>
+						<li><NavLink class="nav-link kulku-sublink" href="network/loginteraction"><Icon Kind="IconKind.ChatDots" Size="14" /> Log Interaction</NavLink></li>
 					</ul>
 				}
 			</li>
