+++ src/Kulku.Web/Kulku.Web.Admin/Components/Network/Components/NetworkCategoryEditModal.razor
@@ -1,4 +1,4 @@
-@using Kulku.Web.Admin.Components.Developer
+@using Kulku.Web.Admin.Components.Network
 
 <ModalDialog Visible="@IsVisible" Title="@(Mode == ModalMode.Create ? "New Category" : "Edit Category")"
 			 Size="ModalSize.Small"
