+++ src/Kulku.Web/Kulku.Web.Admin/Components/Network/Components/NetworkCompanyCard.razor
@@ -14,7 +14,7 @@
 				</div>
 
 				<h6 class="mb-1">
-					<a href="/developer/companies/@Company.CompanyId" class="text-decoration-none">@Company.Name</a>
+					<a href="/network/companies/@Company.CompanyId" class="text-decoration-none">@Company.Name</a>
 				</h6>
 
 				@if (Company.Categories.Count > 0)
