+++ src/Kulku.Web/Kulku.Web.Admin/Components/Network/Components/NetworkContactEditModal.razor
@@ -1,4 +1,4 @@
-@using Kulku.Web.Admin.Components.Developer
+@using Kulku.Web.Admin.Components.Network
 
 <ModalDialog Visible="@IsVisible" Title="@(Mode == ModalMode.Create ? "New Contact" : "Edit Contact")"
 			 Size="ModalSize.Default"
