+++ src/Kulku.Web/Kulku.Web.Admin/Components/Network/Components/NetworkContactEditModal.razor.cs
@@ -2,9 +2,9 @@ using Kulku.Application.Network.Models;
 using Kulku.Web.Admin.Components.Shared;
 using Microsoft.AspNetCore.Components;
 
-namespace Kulku.Web.Admin.Components.Developer.Components;
+namespace Kulku.Web.Admin.Components.Network.Components;
 
-partial class CrmContactEditModal
+partial class NetworkContactEditModal
 {
     [Parameter]
     public bool IsVisible { get; set; }
