+++ src/Kulku.Web/Kulku.Web.Admin/Components/Network/Components/PipelineColumn.razor.cs
@@ -2,7 +2,7 @@ using Kulku.Application.Network.Models;
 using Kulku.Domain.Network;
 using Microsoft.AspNetCore.Components;
 
-namespace Kulku.Web.Admin.Components.Developer.Components;
+namespace Kulku.Web.Admin.Components.Network.Components;
 
 partial class PipelineColumn
 {
