+++ src/Kulku.Web/Kulku.Web.Admin/Components/Network/Pages/LogInteraction.razor
@@ -1,7 +1,7 @@
-@page "/developer/loginteraction"
+@page "/network/loginteraction"
 @using Kulku.Application.Network.Models
 @using Kulku.Domain.Network
-@using Kulku.Web.Admin.Components.Developer.Components
+@using Kulku.Web.Admin.Components.Network.Components
 
 @inject NavigationManager Nav
 
@@ -15,7 +15,7 @@
 			<h1 class="h4 mb-1">Log interaction</h1>
 			<div class="text-muted">Companies are usually created here. Link history when possible.</div>
 		</div>
-		<NavLink class="btn btn-sm btn-outline-secondary" href="developer">Back to dashboard</NavLink>
+		<NavLink class="btn btn-sm btn-outline-secondary" href="network">Back to dashboard</NavLink>
 	</div>
 
 	<div class="row g-3">
@@ -224,7 +224,7 @@
 						{
 							<div class="mb-2">
 								<div class="fw-semibold">@c.Name</div>
-								<div class="text-muted small">@c.Region • @c.Website • @CrmCompanyCard.StageLabel(c.Stage)</div>
+								<div class="text-muted small">@c.Region • @c.Website • @NetworkCompanyCard.StageLabel(c.Stage)</div>
 							</div>
 
 							<div class="mb-2">
