+++ src/Kulku.Web/Kulku.Web.Admin/Components/Network/Pages/LogInteraction.razor.cs
@@ -8,7 +8,7 @@ using Kulku.Domain.Network;
 using Microsoft.AspNetCore.Components;
 using SoulNETLib.Clean.Application.Abstractions.CQRS;
 
-namespace Kulku.Web.Admin.Components.Developer.Pages;
+namespace Kulku.Web.Admin.Components.Network.Pages;
 
 partial class LogInteraction
 {
@@ -239,7 +239,7 @@ partial class LogInteraction
             default
         );
 
-        Nav.NavigateTo("/developer");
+        Nav.NavigateTo("/network");
     }
 
     private static string ChannelLabel(InteractionChannel ch) =>
