+++ src/Kulku.Web/Kulku.Web.Admin/Components/Pages/Error.razor
@@ -8,18 +8,18 @@
 
 @if (ShowRequestId)
 {
-    <p>
-        <strong>Request ID:</strong> <code>@RequestId</code>
-    </p>
+	<p>
+		<strong>Request ID:</strong> <code>@RequestId</code>
+	</p>
 }
 
 <h3>Development Mode</h3>
 <p>
-    Swapping to <strong>Development</strong> environment will display more detailed information about the error that occurred.
+	Swapping to <strong>Development</strong> environment will display more detailed information about the error that occurred.
 </p>
 <p>
-    <strong>The Development environment shouldn't be enabled for deployed applications.</strong>
-    It can result in displaying sensitive information from exceptions to end users.
-    For local debugging, enable the <strong>Development</strong> environment by setting the <strong>ASPNETCORE_ENVIRONMENT</strong> environment variable to <strong>Development</strong>
-    and restarting the app.
+	<strong>The Development environment shouldn't be enabled for deployed applications.</strong>
+	It can result in displaying sensitive information from exceptions to end users.
+	For local debugging, enable the <strong>Development</strong> environment by setting the <strong>ASPNETCORE_ENVIRONMENT</strong> environment variable to <strong>Development</strong>
+	and restarting the app.
 </p>
