+++ src/Kulku.Web/Kulku.Web.Admin/Components/Pages/Home.razor
@@ -9,13 +9,13 @@
 		<div class="container-fluid py-4">
 			<div class="d-flex flex-wrap justify-content-between align-items-start gap-2 mb-4">
 				<div>
-					<h1 class="h3 mb-1">Developer Overview</h1>
+					<h1 class="h3 mb-1">Admin Overview</h1>
 				</div>
 
 				<span class="text-muted">Hello @context.User.Identity?.Name!</span>
 
 				<div class="d-flex flex-wrap gap-2">
-					<NavLink class="btn btn-sm btn-primary" href="developer">Network</NavLink>
+					<NavLink class="btn btn-sm btn-primary" href="network">Network</NavLink>
 					<NavLink class="btn btn-sm btn-outline-secondary" href="ideas">Idea Bank</NavLink>
 				</div>
 			</div>
@@ -119,7 +119,7 @@
 							<div class="mb-3 text-primary">
 								<Icon Kind="IconKind.Lightbulb" Size="36" />
 							</div>
-							<h5 class="card-title">Developer CRM</h5>
+							<h5 class="card-title">Network</h5>
 							<p class="card-text text-muted small">
 								Track companies, interactions, and networking activity
 								to stay on top of your professional outreach.
