+++ src/Kulku.Web/Kulku.Web.Admin/Components/Shared/InfoCard.razor
@@ -1,12 +1,12 @@
 @namespace Kulku.Web.Admin.Components.Shared
 @inherits ComponentBase
 <div class="card shadow-sm h-100">
-    <div class="card-body">
-        <div class="text-uppercase text-muted small mb-1">@Title</div>
-        <div class="d-flex align-items-baseline gap-2">
-            <div class="h2 mb-0">@BadgeCount</div>
-            <span class="badge text-bg-@BadgeColor">@BadgeText</span>
-        </div>
-        <div class="text-muted small mt-2">@Description</div>
-    </div>
+	<div class="card-body">
+		<div class="text-uppercase text-muted small mb-1">@Title</div>
+		<div class="d-flex align-items-baseline gap-2">
+			<div class="h2 mb-0">@BadgeCount</div>
+			<span class="badge text-bg-@BadgeColor">@BadgeText</span>
+		</div>
+		<div class="text-muted small mt-2">@Description</div>
+	</div>
 </div>
